refactor(verus): isolate the trusted base in axioms_t.rs
